+++ manul/src/combinators.rs
@@ -1,7 +1,4 @@
 //! Combinators operating on protocols.
 
 pub mod chain;
-mod markers;
 pub mod misbehave;
-
-pub use markers::{Combinator, CombinatorEntryPoint};