+++ manul/src/protocol/round.rs
@@ -23,7 +23,7 @@ use super::{
 
 /// Possible successful outcomes of [`Round::finalize`].
 #[derive(Debug)]
-pub enum FinalizeOutcome<Id: PartyId, P: Protocol> {
+pub enum FinalizeOutcome<Id: PartyId, P: Protocol<Id>> {
     /// Transition to a new round.
     AnotherRound(BoxedRound<Id, P>),
     /// The protocol reached a result.
@@ -127,12 +127,12 @@ impl RoundId {
 }
 
 /// A distributed protocol.
-pub trait Protocol: 'static {
+pub trait Protocol<Id>: 'static {
     /// The successful result of an execution of this protocol.
     type Result: Debug;
 
     /// An object of this type will be returned when a provable error happens during [`Round::receive_message`].
-    type ProtocolError: ProtocolError;
+    type ProtocolError: ProtocolError<Id>;
 
     /// Returns `Ok(())` if the given direct message cannot be deserialized
     /// assuming it is a direct message from the round `round_id`.
@@ -181,7 +181,7 @@ pub trait Protocol: 'static {
 ///
 /// Provable here means that we can create an evidence object entirely of messages signed by some party,
 /// which, in combination, prove the party's malicious actions.
-pub trait ProtocolError: Debug + Clone + Send + Serialize + for<'de> Deserialize<'de> {
+pub trait ProtocolError<Id>: Debug + Clone + Send + Serialize + for<'de> Deserialize<'de> {
     /// A description of the error that will be included in the generated evidence.
     ///
     /// Make it short and informative.
@@ -234,6 +234,7 @@ pub trait ProtocolError: Debug + Clone + Send + Serialize + for<'de> Deserialize
     fn verify_messages_constitute_error(
         &self,
         deserializer: &Deserializer,
+        guilty_party: &Id,
         shared_randomness: &[u8],
         echo_broadcast: &EchoBroadcast,
         normal_broadcast: &NormalBroadcast,
@@ -247,7 +248,7 @@ pub trait ProtocolError: Debug + Clone + Send + Serialize + for<'de> Deserialize
 
 // A convenience implementation for protocols that don't define any errors.
 // Have to do it for `()`, since `!` is unstable.
-impl ProtocolError for () {
+impl<Id> ProtocolError<Id> for () {
     fn description(&self) -> String {
         panic!("Attempt to use an empty error type in an evidence. This is a bug in the protocol implementation.")
     }
@@ -255,6 +256,7 @@ impl ProtocolError for () {
     fn verify_messages_constitute_error(
         &self,
         _deserializer: &Deserializer,
+        _guilty_party: &Id,
         _shared_randomness: &[u8],
         _echo_broadcast: &EchoBroadcast,
         _normal_broadcast: &NormalBroadcast,
@@ -327,7 +329,7 @@ impl Artifact {
 /// all the others are only reachable throud [`Round::finalize`] by the execution layer.
 pub trait EntryPoint<Id: PartyId> {
     /// The protocol implemented by the round this entry points returns.
-    type Protocol: Protocol;
+    type Protocol: Protocol<Id>;
 
     /// Returns the ID of the round returned by [`Self::make_round`].
     fn entry_round() -> RoundId {
@@ -381,7 +383,7 @@ The way a round will be used by an external caller:
 */
 pub trait Round<Id: PartyId>: 'static + Debug + Send + Sync {
     /// The protocol this round is a part of.
-    type Protocol: Protocol;
+    type Protocol: Protocol<Id>;
 
     /// The round ID.
     ///