+++ manul/src/session/evidence.rs
@@ -76,7 +76,7 @@ impl From<ProtocolValidationError> for EvidenceError {
 /// A self-contained evidence of malicious behavior by a node.
 #[derive_where::derive_where(Debug)]
 #[derive(Clone, Serialize, Deserialize)]
-pub struct Evidence<P: Protocol, SP: SessionParameters> {
+pub struct Evidence<P: Protocol<SP::Verifier>, SP: SessionParameters> {
     guilty_party: SP::Verifier,
     description: String,
     evidence: EvidenceEnum<P, SP>,
@@ -84,7 +84,7 @@ pub struct Evidence<P: Protocol, SP: SessionParameters> {
 
 impl<P, SP> Evidence<P, SP>
 where
-    P: Protocol,
+    P: Protocol<SP::Verifier>,
     SP: SessionParameters,
 {
     pub(crate) fn new_protocol_error(
@@ -193,10 +193,7 @@ where
         Self {
             guilty_party: verifier.clone(),
             description: error.to_string(),
-            evidence: EvidenceEnum::InvalidDirectMessage(InvalidDirectMessageEvidence {
-                direct_message,
-                phantom: core::marker::PhantomData,
-            }),
+            evidence: EvidenceEnum::InvalidDirectMessage(InvalidDirectMessageEvidence(direct_message)),
         }
     }
 
@@ -208,10 +205,7 @@ where
         Self {
             guilty_party: verifier.clone(),
             description: error.to_string(),
-            evidence: EvidenceEnum::InvalidEchoBroadcast(InvalidEchoBroadcastEvidence {
-                echo_broadcast,
-                phantom: core::marker::PhantomData,
-            }),
+            evidence: EvidenceEnum::InvalidEchoBroadcast(InvalidEchoBroadcastEvidence(echo_broadcast)),
         }
     }
 
@@ -223,10 +217,7 @@ where
         Self {
             guilty_party: verifier.clone(),
             description: error.to_string(),
-            evidence: EvidenceEnum::InvalidNormalBroadcast(InvalidNormalBroadcastEvidence {
-                normal_broadcast,
-                phantom: core::marker::PhantomData,
-            }),
+            evidence: EvidenceEnum::InvalidNormalBroadcast(InvalidNormalBroadcastEvidence(normal_broadcast)),
         }
     }
 
@@ -248,9 +239,11 @@ where
         let deserializer = Deserializer::new::<SP::WireFormat>();
         match &self.evidence {
             EvidenceEnum::Protocol(evidence) => evidence.verify::<SP>(&self.guilty_party, &deserializer),
-            EvidenceEnum::InvalidDirectMessage(evidence) => evidence.verify::<SP>(&self.guilty_party, &deserializer),
-            EvidenceEnum::InvalidEchoBroadcast(evidence) => evidence.verify::<SP>(&self.guilty_party, &deserializer),
-            EvidenceEnum::InvalidNormalBroadcast(evidence) => evidence.verify::<SP>(&self.guilty_party, &deserializer),
+            EvidenceEnum::InvalidDirectMessage(evidence) => evidence.verify::<P, SP>(&self.guilty_party, &deserializer),
+            EvidenceEnum::InvalidEchoBroadcast(evidence) => evidence.verify::<P, SP>(&self.guilty_party, &deserializer),
+            EvidenceEnum::InvalidNormalBroadcast(evidence) => {
+                evidence.verify::<P, SP>(&self.guilty_party, &deserializer)
+            }
             EvidenceEnum::InvalidEchoPack(evidence) => evidence.verify(&self.guilty_party, &deserializer),
             EvidenceEnum::MismatchedBroadcasts(evidence) => evidence.verify::<SP>(&self.guilty_party),
         }
@@ -259,11 +252,11 @@ where
 
 #[derive_where::derive_where(Debug)]
 #[derive(Clone, Serialize, Deserialize)]
-enum EvidenceEnum<P: Protocol, SP: SessionParameters> {
-    Protocol(ProtocolEvidence<P>),
-    InvalidDirectMessage(InvalidDirectMessageEvidence<P>),
-    InvalidEchoBroadcast(InvalidEchoBroadcastEvidence<P>),
-    InvalidNormalBroadcast(InvalidNormalBroadcastEvidence<P>),
+enum EvidenceEnum<P: Protocol<SP::Verifier>, SP: SessionParameters> {
+    Protocol(ProtocolEvidence<SP::Verifier, P>),
+    InvalidDirectMessage(InvalidDirectMessageEvidence),
+    InvalidEchoBroadcast(InvalidEchoBroadcastEvidence),
+    InvalidNormalBroadcast(InvalidNormalBroadcastEvidence),
     InvalidEchoPack(InvalidEchoPackEvidence<SP>),
     MismatchedBroadcasts(MismatchedBroadcastsEvidence),
 }
@@ -348,85 +341,67 @@ impl MismatchedBroadcastsEvidence {
     }
 }
 
-#[derive_where::derive_where(Debug)]
-#[derive(Clone, Serialize, Deserialize)]
-pub struct InvalidDirectMessageEvidence<P: Protocol> {
-    direct_message: SignedMessagePart<DirectMessage>,
-    phantom: core::marker::PhantomData<P>,
-}
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InvalidDirectMessageEvidence(SignedMessagePart<DirectMessage>);
 
-impl<P> InvalidDirectMessageEvidence<P>
-where
-    P: Protocol,
-{
-    fn verify<SP>(&self, verifier: &SP::Verifier, deserializer: &Deserializer) -> Result<(), EvidenceError>
+impl InvalidDirectMessageEvidence {
+    fn verify<P, SP>(&self, verifier: &SP::Verifier, deserializer: &Deserializer) -> Result<(), EvidenceError>
     where
+        P: Protocol<SP::Verifier>,
         SP: SessionParameters,
     {
-        let verified_direct_message = self.direct_message.clone().verify::<SP>(verifier)?;
+        let verified_direct_message = self.0.clone().verify::<SP>(verifier)?;
         let payload = verified_direct_message.payload();
 
-        if self.direct_message.metadata().round_id().is_echo() {
+        if self.0.metadata().round_id().is_echo() {
             Ok(EchoRound::<P, SP>::verify_direct_message_is_invalid(payload)?)
         } else {
             Ok(P::verify_direct_message_is_invalid(
                 deserializer,
-                self.direct_message.metadata().round_id(),
+                self.0.metadata().round_id(),
                 payload,
             )?)
         }
     }
 }
 
-#[derive_where::derive_where(Debug)]
-#[derive(Clone, Serialize, Deserialize)]
-pub struct InvalidEchoBroadcastEvidence<P: Protocol> {
-    echo_broadcast: SignedMessagePart<EchoBroadcast>,
-    phantom: core::marker::PhantomData<P>,
-}
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InvalidEchoBroadcastEvidence(SignedMessagePart<EchoBroadcast>);
 
-impl<P> InvalidEchoBroadcastEvidence<P>
-where
-    P: Protocol,
-{
-    fn verify<SP>(&self, verifier: &SP::Verifier, deserializer: &Deserializer) -> Result<(), EvidenceError>
+impl InvalidEchoBroadcastEvidence {
+    fn verify<P, SP>(&self, verifier: &SP::Verifier, deserializer: &Deserializer) -> Result<(), EvidenceError>
     where
+        P: Protocol<SP::Verifier>,
         SP: SessionParameters,
     {
-        let verified_echo_broadcast = self.echo_broadcast.clone().verify::<SP>(verifier)?;
+        let verified_echo_broadcast = self.0.clone().verify::<SP>(verifier)?;
         let payload = verified_echo_broadcast.payload();
 
-        if self.echo_broadcast.metadata().round_id().is_echo() {
+        if self.0.metadata().round_id().is_echo() {
             Ok(EchoRound::<P, SP>::verify_echo_broadcast_is_invalid(payload)?)
         } else {
             Ok(P::verify_echo_broadcast_is_invalid(
                 deserializer,
-                self.echo_broadcast.metadata().round_id(),
+                self.0.metadata().round_id(),
                 payload,
             )?)
         }
     }
 }
 
-#[derive_where::derive_where(Debug)]
-#[derive(Clone, Serialize, Deserialize)]
-pub struct InvalidNormalBroadcastEvidence<P: Protocol> {
-    normal_broadcast: SignedMessagePart<NormalBroadcast>,
-    phantom: core::marker::PhantomData<P>,
-}
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InvalidNormalBroadcastEvidence(SignedMessagePart<NormalBroadcast>);
 
-impl<P> InvalidNormalBroadcastEvidence<P>
-where
-    P: Protocol,
-{
-    fn verify<SP>(&self, verifier: &SP::Verifier, deserializer: &Deserializer) -> Result<(), EvidenceError>
+impl InvalidNormalBroadcastEvidence {
+    fn verify<P, SP>(&self, verifier: &SP::Verifier, deserializer: &Deserializer) -> Result<(), EvidenceError>
     where
+        P: Protocol<SP::Verifier>,
         SP: SessionParameters,
     {
-        let verified_normal_broadcast = self.normal_broadcast.clone().verify::<SP>(verifier)?;
+        let verified_normal_broadcast = self.0.clone().verify::<SP>(verifier)?;
         let payload = verified_normal_broadcast.payload();
 
-        if self.normal_broadcast.metadata().round_id().is_echo() {
+        if self.0.metadata().round_id().is_echo() {
             Ok(EchoRound::<P, SP>::verify_normal_broadcast_is_invalid(
                 deserializer,
                 payload,
@@ -434,7 +409,7 @@ where
         } else {
             Ok(P::verify_normal_broadcast_is_invalid(
                 deserializer,
-                self.normal_broadcast.metadata().round_id(),
+                self.0.metadata().round_id(),
                 payload,
             )?)
         }
@@ -443,7 +418,7 @@ where
 
 #[derive_where::derive_where(Debug)]
 #[derive(Clone, Serialize, Deserialize)]
-struct ProtocolEvidence<P: Protocol> {
+struct ProtocolEvidence<Id, P: Protocol<Id>> {
     error: P::ProtocolError,
     direct_message: SignedMessagePart<DirectMessage>,
     echo_broadcast: SignedMessagePart<EchoBroadcast>,
@@ -454,13 +429,13 @@ struct ProtocolEvidence<P: Protocol> {
     combined_echos: SerializableMap<RoundId, SignedMessagePart<NormalBroadcast>>,
 }
 
-impl<P> ProtocolEvidence<P>
+impl<Id, P> ProtocolEvidence<Id, P>
 where
-    P: Protocol,
+    P: Protocol<Id>,
 {
     fn verify<SP>(&self, verifier: &SP::Verifier, deserializer: &Deserializer) -> Result<(), EvidenceError>
     where
-        SP: SessionParameters,
+        SP: SessionParameters<Verifier = Id>,
     {
         let session_id = self.direct_message.metadata().session_id();
 
@@ -549,6 +524,7 @@ where
 
         Ok(self.error.verify_messages_constitute_error(
             deserializer,
+            verifier,
             session_id.as_ref(),
             &verified_echo_broadcast,
             &verified_normal_broadcast,