+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `protocol::digest` re-export ([#75]).
 - `digest` and `signature` are now re-exported from the top level instead of `session` ([#75]).
 - `ProtocolError::verify_messages_constitute_error()` takes a new `shared_randomness` argument. ([#76])
+- `Protocol` and `ProtocolError` are now generic over `Id`. ([#76])
+- `ProtocolError::verify_messages_constitute_error()` takes a new `guilty_party` argument. ([#76])
+- `Combinator`/`CombinatorEntryPoint` removed in favor of a single `ChainedMarker` trait. ([#76])
 
 
 [#75]: https://github.com/entropyxyz/manul/pull/75